🎨 メッセージ表示部分を別コンポーネントに切り分け
保守性を上げるため
