+++ src/pages/parking.tsx
@@ -5,27 +5,13 @@ import data from "@/data/data.json";
 import { Box, Tag, Text } from "@chakra-ui/react";
 import { ChangeEvent, useState } from "react";
 import { Parking } from "@/types/Parking";
+import ParkingMessage from "@/components/ParkingMessage";
 
 type Props = InferGetServerSidePropsType<typeof getServerSideProps>;
 
 const Parking: NextPage<Props> = ({ parking }) => {
   const [selectedParking, setSelectedParking] = useState<Parking>(parking);
   const parkings = data.parkings;
-
-  const suggestMessage = (now: Date, parking: Parking) => {
-    const parkingOpenTime = new Date(parking.openAt);
-    const parkingCloseTime = new Date(parking.closeAt);
-
-    if (now.getTime() < parkingOpenTime.getTime()) {
-      // 開場前
-      return `開場前です。\n${parkingOpenTime.getHours()}時に開場します。`;
-    } else if (now.getTime() > parkingCloseTime.getTime()) {
-      // 閉場後
-      return "閉場しました。";
-    } else {
-      return "開場中";
-    }
-  };
 
   const handleChange = (e: ChangeEvent<HTMLSelectElement>) => {
     // 渡された key から駐車場を特定する
@@ -44,14 +30,7 @@ const Parking: NextPage<Props> = ({ parking }) => {
           ))}
         </Select>
       </Box>
-      <Box boxShadow="xs" py={3} px={5}>
-        <Tag size="sm" bgColor="blue.300" fontWeight="bold" color="white" mb={1}>
-          開場前
-        </Tag>
-        <Text color="black" fontSize="sm">
-          {suggestMessage(new Date(), selectedParking)}
-        </Text>
-      </Box>
+      <ParkingMessage parking={selectedParking} />
     </Container>
   );
 };
